[#152656203] Changes to block customers with no invoices and readd work to work with filters

+++ KeithLink/KeithLink.Svc.Core/Models/Invoices/InvoiceCustomer.cs
@@ -20,7 +20,7 @@ namespace KeithLink.Svc.Core.Models.Invoices
         public decimal TotalAmountDue { get; set; }
 
         [DataMember(Name = "numberInvoices")]
-        public int NumberInvoices { get; set; }
+        public long NumberInvoices { get; set; }
 
         [DataMember(Name = "customerNumber")]
         public string CustomerNumber { get; set; }
